flowey build-igvm: fix --with-perf-tools

+++ openhcl/update-rootfs.py
@@ -11,7 +11,7 @@ import subprocess
 import tempfile
 import time
 
-from gen_init_ramfs import create_cpio_from_config, create_cpio_from_dir
+from gen_init_ramfs import create_cpio_from_config, create_cpio_from_dir, create_cpio_from_file
 from typing import List
 
 package_id = "Microsoft.HCL.Kernel"
@@ -88,7 +88,7 @@ def generate_build_info(path: str):
 def process(temp_dir: str, underhill_path: str, kernel_path: str,
             build_info: str, rootfs_config_path: List[str],
             updated_initramfs_path: str, additional_layers: List[str],
-            additional_dirs: List[str]):
+            additional_dirs: List[str], add_perf: bool):
     align = lambda x, boundary: (x + boundary-1) & ~(boundary-1)
 
     eprint("Building the initial root fs")
@@ -111,6 +111,17 @@ def process(temp_dir: str, underhill_path: str, kernel_path: str,
     for dir_name in additional_dirs:
         temp_file_name = os.path.join(temp_dir, "add_dir.cpio.gz")
         create_cpio_from_dir(dir_name, temp_file_name, 'gzip')
+        append_file(underhill_cpio_gz_file_name, temp_file_name)
+        os.unlink(temp_file_name)
+
+    if add_perf:
+        temp_file_name = os.path.join(temp_dir, "perf_tools.cpio.gz")
+        mode = '0755'
+        perf_tool_bin_path = os.path.join(kernel_path, "build", "native", "bin", os.environ["OPENHCL_KERNEL_ARCH"], "tools", "perf", "bin")
+        create_cpio_from_file(os.path.join(perf_tool_bin_path, "perf"), "/usr/bin/perf", mode, temp_file_name, 'gzip')
+        append_file(underhill_cpio_gz_file_name, temp_file_name)
+        os.unlink(temp_file_name)
+        create_cpio_from_file(os.path.join(perf_tool_bin_path, "trace"), "/usr/bin/trace", mode, temp_file_name, 'gzip')
         append_file(underhill_cpio_gz_file_name, temp_file_name)
         os.unlink(temp_file_name)
 
@@ -151,6 +162,7 @@ def main():
     parser.add_argument('--build_info', help='Path to the file with build information')
     parser.add_argument('--rootfs-config', action='append', help='Configuration file for the root filesystem')
     parser.add_argument('--layer', action='append', help='Adds a custom layer file.')
+    parser.add_argument('--perf', action='store_true', help='Adds perf and trace tools')
     parser.add_argument('--add-dir', action='append', help='Adds a directory.')
     parser.set_defaults(layer=[], add_dir=[], rootfs_config=[])
 
@@ -223,7 +235,7 @@ def main():
             str(temp_dir),
             underhill_path, kernel_path, build_info, args.rootfs_config,
             updated_initramfs_path, additional_layers,
-            additional_dirs)
+            additional_dirs, args.perf)
 
 
 if __name__ == '__main__':
